requirements.txt
pyaml==18.11.0
PyYAML==4.2b1
six==1.12.0
websocket-client==0.54.0
